Builds OpenDSSCmd from a common code base using Lazarus 1.6.4 with FPC 3.0.2. Locally, CmdForms.pas replaces all of the forms and dialogs, while KLUSolve.pas replaces the one for Delphi under ../Common.
Intermediate outputs go to a created units subdirectory. Lib files for Win64, Mac and Linux are version controlled in the lib subdirectory. The executable output goes to the test subdirectory, which contains a few IEEE 13-bus test files. You need to copy the platform-dependent file from ./lib to ./test in order to run tests. 

git-svn-id: https://svn.code.sf.net/p/electricdss/code@1925 d8739450-1e93-4ef4-a0af-c327d92816ff