OpenDSSC/Plotting: fix and enable `Visualize`, `DI_Plot`, and `YearlyCurves` commands; allow generating DSV files when `AllowForms=false`.
DSV files are useful even without forms for testing and validation, but also to allow general automation -- e.g. run a DSS script unattended, generating the plot DSV files for later inspection. We recommend adding a similar change to the Delphi version to make it easier to compare the outputs.

git-svn-id: https://svn.code.sf.net/p/electricdss/code/trunk@3883 d8739450-1e93-4ef4-a0af-c327d92816ff