OpenDSSC/DDLL: Add new function `DSSDisposeString` to free the memory used by the strings from various functions of the ODD API.
We need this to avoid memory leaks when using the DirectDLL API.

It is not safe to call plain `free()` from another library, and the strings are allocated with `new char[...]` instead of plain `malloc`.
Besides that, not every language and environment provides trivial access to the `free` function from the C runtime anyway.

git-svn-id: https://svn.code.sf.net/p/electricdss/code/trunk@3894 d8739450-1e93-4ef4-a0af-c327d92816ff