Updates to Line.pas, LineConstants.pas and LineGeometry.pas. Adding three properties ('HeightOffset', 'HeightUnit', 'conductors'). The first two are intended to add an offset to the Y coordinates of conductors in the cross-section plane for lines using geometries or spacings. Helps to better map data to/from commercial tools where lines using geometries/spacings have their own average height property independent from the cross-section plane coordinates.
The third property, "conductors", is an array added on top of the existing "wires" "cncables", "tscables". The intention of this one is to allow defining the array of conductors using conductor type and conductor name in the same array: 

conductors=[cndata.CableConcentricNeutral, None, tsdata.CableTapeShield, wiredata.CableSeparateNeutral]

Lastly, this update allows to define empty spaces in the "conductors" array (as well as the existing "wires" "cncables", "tscables"). Empty spaces are set to "None" (as in the example above). This helps avoid the need to define different spacings for each different combination of existing/missing conductors in the conductor arrangement (a cross-arm for instance).



 

git-svn-id: https://svn.code.sf.net/p/electricdss/code@3897 d8739450-1e93-4ef4-a0af-c327d92816ff