Updates to Delphi and Cpp versions to allow for the following:
(1) Added "conductors" array attribute (Prop #20) to LineGeometry to be used along with a LineSpacing similar to the recently added feature of Line objects (conductors=[cndata.cncablename, cndata.cncablename, tsdata.tscablename, wiredata.wirename]; "None" should be used to define unused positions in the referenced LineSpacing).

(2) Added Properties to LineSpacing: "detailed" (#6), "EqDistPhPh"(#7), "EqDistPhN"(#8), "AvgPhaseHeight" (#9), and "AvgNeutralHeight" (#10). "Detailed" is Yes/True by default which disregards the other properties (expecting the usual cross-sectional X and H arrays or coordinates). This is intended to model Equivalent Phase to Phase and Phase to Neutral distances which is typically available in commercial software platforms. The resulting series impedance and shunt admittance matrices are symmetrical and the diagonal elements are identical. 

These additions require updates to LineConstants and CableConstants to account for the usage of either the detailed cross-sectional coordinates or the equivalent ones.

git-svn-id: https://svn.code.sf.net/p/electricdss/code@3910 d8739450-1e93-4ef4-a0af-c327d92816ff